+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.0 (beta)
 
 - Removed automatic call to `init()` when creating the first logging record.
+- Removed `reset=True` option for `log.init()`. Use `log.reset()` instead.
 
 # 1.6 (2020-05-24)